supabase-vue-3/requirements.txt
gradio
openai
supabase
python-dotenv
mangum